requirements.txt
fastapi~=0.110.0
uvicorn~=0.27.1
transformers~=4.36.2
torch~=2.1.2
Pillow~=9.5.0